# Conflicts:
#	erp-server/erp-server-tms/src/main/resources/mapper/LogisticsBillMapper.xml

+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/order/BuyerBeanX.java
@@ -15,7 +15,7 @@ public class BuyerBeanX {
      */
 
     @JsonProperty("id")
-    private int fid;
+    private long fid;
     @JsonProperty("nickname")
     private String nickname;
     @JsonProperty("last_name")
